Move doMove inside of Piece
doMove is only ever called from inside of Piece
